requirements.txt
numpy
coloredlogs
watchtower
exceptiongroup
python-dotenv
flake8
pypdfium2
boto3
pandas
black
natsort
fuzzywuzzy
pytest
openai
pdfplumber
jsonschema
langchain
langgraph
httpx
aiofiles
neo4j==4.4.11
requests-aws4auth
langchain-community
google-cloud-vision
pymupdf
pydantic
pydantic-settings
python-multipart
vllm
pylint
rich
pycocotools
opencv-python-headless
tomli
dill
PyPDF2
langchain-openai
fastapi
uvicorn
types-requests
azure-core==1.32.0
azure-ai-documentintelligence==1.0.0
azure-ai-formrecognizer==3.3.3
azure-common==1.1.28
agentic-doc

# Development packages
ipykernel
notebook
landingai-ade
